rdk: CB2 Fix WPE crash when youtube exit [1/1]
PD#SWPL-149515

Problem:
"Libgstamlsvpext. so" is dynamically loaded by cobalt.
When so is loaded, a TimerCallback() callback function is added to glib
in the initialization function.
However, when so is unloaded, the deinitialization function returns midway
and the TimerCallback() callback function is not deleted.
This Causing other modules to access empty function pointers
in the "g_main_loop_run()" function.

Solution:
When an error occurs during the deinitialization function,
do not exit to ensure resource release

Verify:
T5W-AT301

Signed-off-by: zhenyu.gao <zhenyu.gao@amlogic.com>
Change-Id: Ie116870f2d3f916af2f6ca99249e4195c99112d7